fix: resolve error handling lint errors
