+++ backend/go.mod
@@ -3,3 +3,8 @@ module backend
 go 1.24
 
 require github.com/golang-jwt/jwt/v5 v5.2.2
+
+require (
+	go.uber.org/multierr v1.10.0 // indirect
+	go.uber.org/zap v1.27.0 // indirect
+)
